Convert all arrays to float32

+++ src/net.py
@@ -634,11 +634,11 @@ class Net(object):
 
                 # img1 + matches + sparse + gt flow
                 elif len(path_inputs) == 4 and input_type == 'image_matches':
-                    frame_0 = np.array(imread(path_inputs[0]))
+                    frame_0 = np.array(imread(path_inputs[0])).astype(np.float32)
                     frame_1 = None
-                    matches_0 = np.array(imread(path_inputs[1]))
-                    sparse_flow_0 = np.array(read_flow(path_inputs[2]))
-                    gt_flow_0 = np.array(read_flow(path_inputs[3]))
+                    matches_0 = np.array(imread(path_inputs[1])).astype(np.float32)
+                    sparse_flow_0 = np.array(read_flow(path_inputs[2])).astype(np.float32)
+                    gt_flow_0 = np.array(read_flow(path_inputs[3])).astype(np.float32)
                     if compute_metrics:
                         # Must define optional masks as None
                         occ_mask_0 = None
